requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-extra-fields==3.7.0
filetype==1.2.0
oracledb==2.1.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
sqlparse==0.4.4
tzdata==2024.1
